--- src/main/java/com/practice/Skilltest/SecurityConfig.java
@@ -1,9 +1,0 @@
-package com.practice.Skilltest;
-
-public class SecurityConfig {
-
-
-
-
-
-}
