+++ src/main/resources/mybatis/mapper/UserMapper.xml
@@ -2,7 +2,9 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
+<!-- 유저 로그인 XML -->
 <mapper namespace="com.practice.Skilltest.user.dao.UserLoginDao">
+
